moved types declaration file to root
