+++ src/app/defines.ts
@@ -1,4 +1,4 @@
-import { Environment } from '../types';
+import { Environment } from '../../types';
 
 export const ENV_DEVELOPMENT: Environment = 'dev';
 export const ENV_PRODUCTION: Environment = 'prod';
