+++ src/app/services/twigService.ts
@@ -1,5 +1,5 @@
 import Twig from 'twig';
-import { TwigConfiguration } from '../../types';
+import { TwigConfiguration } from '../../../types';
 import { getNestedValue } from '../../utils/getNestedValue';
 
 const sort = (a: any, b: any): number => {
