+++ tsconfig.json
@@ -68,7 +68,7 @@
     "forceConsistentCasingInFileNames": true  /* Disallow inconsistently-cased references to the same file. */
   },
   "include": [
-    "src"
+    "src", "types.d.ts"
   ],
   "exclude": [
     "**/*.spec.ts"
